# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
...
  user.Login:
    properties:
      long_token:
        description: If true, the token returned will be valid a lot longer than default.
          Useful for "remember me" style logins.
        type: boolean
      password:
        description: The password for the user.
        type: string
      totp_passcode:
        description: The totp passcode of a user. Only needs to be provided when enabled.
        type: string
      username:
        description: The username used to log in.
        type: string
    type: object
...
  user.TOTP:
    properties:
      enabled:
        description: The totp entry will only be enabled after the user verified they
          have a working totp setup.
        type: boolean
      secret:
        type: string
      url:
        description: The totp url used to be able to enroll the user later
        type: string
    type: object
  user.TOTPPasscode:
    properties:
      passcode:
        type: string
    type: object
...
  web.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact:
    email: hello@vikunja.io
    name: General Vikunja contact
    url: https://vikunja.io/contact/
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. Make sure to check the [error docs](https://vikunja.io/docs/errors/) in Vikunja's documentation for a full list of available error codes.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

...

    **BasicAuth:** Only used when requesting tasks via CalDAV.
...
  license:
...
paths:
...
  /user/settings/totp:
    get:
      consumes:
      - application/json
      description: Returns the current user totp setting or an error if it is not
        enabled.
      produces:
      - application/json
      responses:
        "200":
          description: The totp settings.
          schema:
            $ref: '#/definitions/user.TOTP'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Totp setting for the current user
      tags:
      - user
  /user/settings/totp/disable:
    post:
      consumes:
      - application/json
      description: Disables any totp settings for the current user.
      parameters:
      - description: The current user's password (only password is enough).
        in: body
        name: totp
        required: true
        schema:
          $ref: '#/definitions/user.Login'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully disabled
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Disable totp settings
      tags:
      - user
  /user/settings/totp/enable:
    post:
      consumes:
      - application/json
      description: Enables a previously enrolled totp setting by providing a totp
        passcode.
      parameters:
      - description: The totp passcode.
        in: body
        name: totp
        required: true
        schema:
          $ref: '#/definitions/user.TOTPPasscode'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully enabled
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "412":
          description: TOTP is not enrolled.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Enable a previously enrolled totp setting.
      tags:
      - user
  /user/settings/totp/enroll:
    post:
      consumes:
      - application/json
      description: Creates an initial setup for the user in the db. After this step,
        the user needs to verify they have a working totp setup with the "enable totp"
        endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.TOTP'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Enroll a user into totp
      tags:
      - user
  /user/settings/totp/qrcode:
    get:
      consumes:
      - application/json
      description: Returns a qr code for easier setup at end user's devices.
      produces:
      - application/json
      responses:
        "200":
          description: The qr code as jpeg image
          schema:
            type: file
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Totp QR Code
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"